优化

+++ app/request/user.go
@@ -42,9 +42,9 @@ type User struct {
 	Username string `json:"username" form:"username" validate:"required" label:"用户名"`
 	FullName string `json:"fullName" validate:"required" label:"姓名"`
 	Nickname string `json:"nickname" validate:"required" label:"昵称"`
-	Gender   int    `json:"gender" validate:"required|int" label:"性别"`
+	Gender   int64  `json:"gender" validate:"required|int" label:"性别"`
 	Password string `json:"password" validate:"required" label:"密码"`
-	Age      int    `json:"age" validate:"int" label:"年龄"`
+	Age      int64  `json:"age" validate:"int" label:"年龄"`
 	PageListValidate
 }
